+++ app/server/api_test/api_read_test/product_test.go
@@ -4,7 +4,7 @@ package api_read_test
 
 import (
 	"encoding/json"
-	"github/code-kakitai/code-kakitai/server/utils"
+	"github/code-kakitai/code-kakitai/server/api_test/test_utils"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -90,7 +90,7 @@ func TestProduct_GetProducts_With_Goldie(t *testing.T) {
 				goldie.WithNameSuffix(".golden.json"),
 				goldie.WithFixtureDir("testdata/product_test"),
 			)
-			g.Assert(t, t.Name(), utils.FormatJSON(t, w.Body.Bytes()))
+			g.Assert(t, t.Name(), test_utils.FormatJSON(t, w.Body.Bytes()))
 		})
 	}
 }
