+++ app/server/api_test/test_utils/common.go
@@ -1,4 +1,4 @@
-package utils
+package test_utils
 
 import (
 	"bytes"
