+++ app/server/api_test/test_utils/goldie_options.go
@@ -1,4 +1,4 @@
-package utils
+package test_utils
 
 import (
 	"encoding/json"
